+++ tests/test_profobuf.py
@@ -1,7 +1,7 @@
 from http import HTTPStatus
 
 import pytest
-from httpx import ConnectTimeout
+from httpx import ConnectTimeout, HTTPStatusError
 from pytest_httpx import HTTPXMock
 from pytest_mock import MockerFixture
 
@@ -32,7 +32,7 @@ class TestProtobuf:
 
     @pytest.mark.asyncio
     async def test__async_get_wrong_password(self, httpx_mock: HTTPXMock, mock_protobuf: StubProtobuf):
-        httpx_mock.add_response(status_code=401)
+        httpx_mock.add_response(status_code=HTTPStatus.UNAUTHORIZED)
         with pytest.raises(DevicePasswordProtected):
             await mock_protobuf._async_get("LedSettingsGet")
 
@@ -40,6 +40,12 @@ class TestProtobuf:
     async def test__async_get_device_unavailable(self, httpx_mock: HTTPXMock, mock_protobuf: StubProtobuf):
         httpx_mock.add_exception(ConnectTimeout("ConnectTimeout"))
         with pytest.raises(DeviceUnavailable):
+            await mock_protobuf._async_get("LedSettingsGet")
+
+    @pytest.mark.asyncio
+    async def test__async_get_unknown_error(self, httpx_mock: HTTPXMock, mock_protobuf: StubProtobuf):
+        httpx_mock.add_response(status_code=HTTPStatus.SERVICE_UNAVAILABLE)
+        with pytest.raises(HTTPStatusError):
             await mock_protobuf._async_get("LedSettingsGet")
 
     def test__message_to_dict(self, mocker: MockerFixture, mock_protobuf: StubProtobuf):
@@ -55,7 +61,7 @@ class TestProtobuf:
 
     @pytest.mark.asyncio
     async def test__async_post_wrong_password(self, httpx_mock: HTTPXMock, mock_protobuf: StubProtobuf):
-        httpx_mock.add_response(status_code=401)
+        httpx_mock.add_response(status_code=HTTPStatus.UNAUTHORIZED)
         with pytest.raises(DevicePasswordProtected):
             await mock_protobuf._async_post("LedSettingsGet", b"")
 
@@ -63,6 +69,12 @@ class TestProtobuf:
     async def test__async_post_device_unavailable(self, httpx_mock: HTTPXMock, mock_protobuf: StubProtobuf):
         httpx_mock.add_exception(ConnectTimeout("ConnectTimeout"))
         with pytest.raises(DeviceUnavailable):
+            await mock_protobuf._async_post("LedSettingsGet", b"")
+
+    @pytest.mark.asyncio
+    async def test__async_post_unknown_error(self, httpx_mock: HTTPXMock, mock_protobuf: StubProtobuf):
+        httpx_mock.add_response(status_code=HTTPStatus.SERVICE_UNAVAILABLE)
+        with pytest.raises(HTTPStatusError):
             await mock_protobuf._async_post("LedSettingsGet", b"")
 
     @pytest.mark.asyncio