+++ docs/CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+#### Device API
+
+- Specify a duration for the guest wifi
+- Start WPS clone mode
+
 #### PLC API
 
 - Start pairing mode